Use BigInt instead of BigInteger

+++ protocol/src/main/scala/org/alephium/protocol/model/Block.scala
@@ -1,7 +1,5 @@
 package org.alephium.protocol.model
 
-import java.math.BigInteger
-
 import org.alephium.crypto.{Keccak256, WithKeccak256}
 import org.alephium.serde.Serde
 
@@ -13,7 +11,7 @@ case class Block(blockHeader: BlockHeader, transactions: Seq[Transaction])
 object Block {
   implicit val serde: Serde[Block] = Serde.forProduct2(apply, b => (b.blockHeader, b.transactions))
 
-  def from(blockDeps: Seq[Keccak256], transactions: Seq[Transaction], nonce: BigInteger): Block = {
+  def from(blockDeps: Seq[Keccak256], transactions: Seq[Transaction], nonce: BigInt): Block = {
     // TODO: validate all the block dependencies; the first block dep should be previous block in the same chain
     val txsHash     = Keccak256.hash(transactions)
     val timestamp   = System.currentTimeMillis()
